T51: refactor settings based on feature not screens
