+++ app/src/main/java/com/gals/prayertimes/settings/calendar/PrayersCalendarViewModel.kt
@@ -1,4 +1,4 @@
-package com.gals.prayertimes.settings
+package com.gals.prayertimes.settings.calendar
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -7,7 +7,7 @@ import com.gals.prayertimes.common.ConnectivityException
 import com.gals.prayertimes.common.ServerException
 import com.gals.prayertimes.common.mappers.toPrayer
 import com.gals.prayertimes.repository.PrayersRepository
-import com.gals.prayertimes.settings.model.PrayersCalendarUiState
+import com.gals.prayertimes.settings.calendar.model.PrayersCalendarUiState
 import com.gals.prayertimes.utils.DateFormatter
 import com.gals.prayertimes.utils.ResourceProvider
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -20,27 +20,28 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-
 @HiltViewModel
 class PrayersCalendarViewModel @Inject constructor(
     private val repository: PrayersRepository,
     private val dateFormatter: DateFormatter,
     private val resourceProvider: ResourceProvider
 ) : ViewModel() {
-    /*TODO: tracking, ads, tests, regions for VM, add outlined cards to notification screen*/
+    /*TODO: tracking, ads, tests, regions for VM, add outlined cards to notification screen, thinks first load case form locale, extract outlinedCard implementation for notification screen also*/
     private val _uiState = MutableStateFlow<PrayersCalendarUiState>(PrayersCalendarUiState.Loading)
     val uiState: StateFlow<PrayersCalendarUiState> = _uiState.asStateFlow()
-
     fun fetchPrayersForDate(selectedDateMillis: Long) {
         _uiState.update { PrayersCalendarUiState.Loading }
-        val selectedDate = dateFormatter.formatMillisToDateString(selectedDateMillis)
         viewModelScope.launch {
-            repository.getRemotePrayer(selectedDate)
-                .map { it.toPrayer(resourceProvider, dateFormatter) }
-                .catch { cause ->
-                    _uiState.update { cause.toUiError() }
-                }
-                .collect { uiPrayer -> _uiState.update { PrayersCalendarUiState.Success(uiPrayer) } }
+            dateFormatter.formatMillisToDateString(selectedDateMillis).let { selectedDate ->
+                repository.getRemotePrayer(selectedDate)
+                    .catch { cause ->
+                        _uiState.update { cause.toUiError() }
+                    }
+                    .map { it.toPrayer(resourceProvider, dateFormatter) }
+                    .collect { uiPrayer ->
+                        _uiState.update { PrayersCalendarUiState.Success(uiPrayer) }
+                    }
+            }
         }
     }
