+++ app/src/main/java/com/gals/prayertimes/settings/calendar/tracker/PrayerCalendarTracker.kt
@@ -1,0 +1,5 @@
+package com.gals.prayertimes.settings.calendar.tracker
+
+interface PrayerCalendarTracker {
+
+}
