+++ app/src/main/java/com/gals/prayertimes/settings/notification/tracker/NotificationSettingsTracker.kt
@@ -1,4 +1,4 @@
-package com.gals.prayertimes.settings.tracker
+package com.gals.prayertimes.settings.notification.tracker
 
 interface NotificationSettingsTracker {
     fun submitSettings(params: Map<String, Any>)
